app/requirements.txt
blinker==1.9.0
boto3==1.35.71
botocore==1.35.71
click==8.1.7
Flask==3.1.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
s3transfer==0.10.4
six==1.16.0
urllib3==1.26.20
Werkzeug==3.1.3
zipp==3.21.0
